criando pagina de filtro para busca de informações

+++ divulgar/views.py
@@ -56,6 +56,12 @@ def seus_pets(request):
     
 def remover_pet(request, id):
     pet = Pet.objects.get(id=id)
+    
+    if not pet.usuario == request.user:
+        messages.add_message(request, constants.ERROR, 'Esse PET não é seu!\n')
+        return redirect('/divulgar/seus_pets')
+    
+
     pet.delete()
     
     messages.add_message(request, constants.SUCCESS, "Pet removido com sucesso")
